+++ oap-server/server-alarm-plugin/src/main/java/org/apache/skywalking/oap/server/core/alarm/provider/dingtalk/DingtalkHookCallback.java
@@ -134,6 +134,7 @@ public class DingtalkHookCallback implements AlarmCallback {
      * Send alarm message to remote endpoint
      */
     private void sendAlarmMessage(CloseableHttpClient httpClient, String url, String requestBody) {
+        CloseableHttpResponse httpResponse = null;
         try {
             HttpPost post = new HttpPost(url);
             post.setConfig(requestConfig);
@@ -141,7 +142,7 @@ public class DingtalkHookCallback implements AlarmCallback {
             post.setHeader(HttpHeaders.CONTENT_TYPE, HttpHeaderValues.APPLICATION_JSON.toString());
             StringEntity entity = new StringEntity(requestBody, ContentType.APPLICATION_JSON);
             post.setEntity(entity);
-            CloseableHttpResponse httpResponse = httpClient.execute(post);
+            httpResponse = httpClient.execute(post);
             StatusLine statusLine = httpResponse.getStatusLine();
             if (statusLine != null && statusLine.getStatusCode() != HttpStatus.SC_OK) {
                 log.error("send dingtalk alarm to {} failure. Response code: {}, Response content: {}", url, statusLine.getStatusCode(),
@@ -149,6 +150,15 @@ public class DingtalkHookCallback implements AlarmCallback {
             }
         } catch (Throwable e) {
             log.error("send dingtalk alarm to {} failure.", url, e);
+        } finally {
+            if (httpResponse != null) {
+                try {
+                    httpResponse.close();
+                } catch (IOException e) {
+                    log.error(e.getMessage(), e);
+                }
+
+            }
         }
     }
 }